requirements.txt
discord.py
python-digitalocean
